requirements.txt
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
clip @ https://github.com/openai/CLIP/archive/d50d76daa670286dd6cacf3bcd80b5e4823fc8e1.zip#sha256=b5842c25da441d6c581b53a5c60e0c2127ebafe0f746f8e15561a006c6c3be6a
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
ffmpy==0.4.0
filelock==3.15.4
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.53.1
fsspec==2024.6.1
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
gradio==4.40.0
gradio_client==1.2.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
imageio==2.34.2
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lapx==0.5.9.post1
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
open_clip_torch @ https://github.com/mlfoundations/open_clip/archive/bb6e834e9c70d9c27d0dc3ecedeebeaeb1ffad6b.zip#sha256=a594e737394d4463872151a1e84ab5fa502a4f72686943526b663fa939d786b1
openai==1.51.2
opencv-python==4.10.0.84
orjson==3.10.6
packaging==24.1
pandas==2.2.2
Pillow==9.2.0
protobuf==3.20.0
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==16.1.0
pycocotools==2.0.8
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytubefix==6.1.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
ruff==0.5.5
scikit-image==0.24.0
scipy==1.14.0
seaborn==0.13.2
-e git+https://github.com/z-x-yang/Segment-and-Track-Anything.git@228b6f6ba6961e3618c9f27684746efcaed99a74#egg=segment_anything&subdirectory=sam
semantic-version==2.10.0
sentencepiece==0.2.0
setuptools==70.3.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
streamlit==1.36.0
sympy==1.13.0
tenacity==8.5.0
thop==0.1.1.post2209072238
tifffile==2024.7.24
timm==0.4.5
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.56
ultralytics-thop==2.0.0
urllib3==2.2.2
uvicorn==0.30.4
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.4
wget==3.2
